Run unit tests with pytest on Python >= 3.10
