+++ ci/nosetests.sh
@@ -5,18 +5,18 @@ set -u
 EXIT_CODE=0
 
 echo "========================================="
-echo "nosetests -a '!requires_sudo' output"
+echo "****     nosetests non-sudo tests    ****"
 echo "========================================="
-nosetests -a '!requires_sudo' tests $NOSEOPTS || EXIT_CODE=$(($EXIT_CODE || $?))
+nosetests --ignore-files test_cgroupconfigurator_sudo.py tests $NOSEOPTS || EXIT_CODE=$(($EXIT_CODE || $?))
 echo EXIT_CODE no_sudo nosetests = $EXIT_CODE
 
 [[ -f .coverage ]] && \
     sudo mv .coverage coverage.$(uuidgen).no_sudo.data
 
 echo "========================================="
-echo "nosetests -a 'requires_sudo' output"
+echo "****      nosetests sudo tests       ****"
 echo "========================================="
-sudo env "PATH=$PATH" nosetests -a 'requires_sudo' tests $NOSEOPTS || EXIT_CODE=$(($EXIT_CODE || $?))
+sudo env "PATH=$PATH" nosetests tests/ga/test_cgroupconfigurator_sudo.py $NOSEOPTS || EXIT_CODE=$(($EXIT_CODE || $?))
 echo EXIT_CODE with_sudo nosetests = $EXIT_CODE
 
 [[ -f .coverage ]] && \
