+++ test-requirements.txt
@@ -3,8 +3,9 @@ mock==2.0.0; python_version == '2.6'
 mock==3.0.5; python_version >= '2.7' and python_version <= '3.5'
 mock==4.0.2; python_version >= '3.6'
 distro; python_version >= '3.8'
-nose
-nose-timer; python_version >= '2.7'
+nose; python_version <= '3.9'
+nose-timer; python_version >= '2.7' and python_version <= '3.9'
+pytest; python_version >= '3.10'
 
 # Pinning the wrapt requirement to 1.12.0 due to the bug - https://github.com/GrahamDumpleton/wrapt/issues/188
 wrapt==1.12.0; python_version > '2.6' and python_version < '3.6'
